Extract backup file format from FileBackupAgent
