+++ fdbclient/FileBackupAgent.cpp
@@ -44,6 +44,7 @@
 #include "fdbclient/Status.h"
 #include "fdbclient/SystemData.h"
 #include "fdbclient/TaskBucket.h"
+#include "FileBackupAgentFileFormat.h"
 #include "flow/network.h"
 #include "flow/Trace.h"
 #include "flow/Util.h"
@@ -1497,367 +1498,10 @@ FileBackupAgent::FileBackupAgent()
 
 namespace fileBackup {
 
-// Return a block of contiguous padding bytes, growing if needed.
-Value makePadding(int size) {
-	static Value pad;
-	if (pad.size() < size) {
-		pad = makeString(size);
-		memset(mutateString(pad), '\xff', pad.size());
-	}
-
-	return pad.substr(0, size);
-}
-
-struct IRangeFileWriter {
-public:
-	virtual Future<Void> padEnd(bool final) = 0;
-
-	virtual Future<Void> writeKV(Key k, Value v) = 0;
-
-	virtual Future<Void> writeKey(Key k) = 0;
-
-	virtual Future<Void> finish() = 0;
-
-	virtual ~IRangeFileWriter() = default;
-};
-
-// File Format handlers.
-// Both Range and Log formats are designed to be readable starting at any BACKUP_RANGEFILE_BLOCK_SIZE boundary
-// so they can be read in parallel.
-//
-// Writer instances must be kept alive while any member actors are in progress.
-//
-// RangeFileWriter must be used as follows:
-//   1 - writeKey(key) the queried key range begin
-//   2 - writeKV(k, v) each kv pair to restore
-//   3 - writeKey(key) the queried key range end
-//	 4 - finish()
-//
-// RangeFileWriter will insert the required padding, header, and extra
-// end/begin keys around the 1MB boundaries as needed.
-//
-// Example:
-//   The range a-z is queries and returns c-j which covers 3 blocks.
-//   The client code writes keys in this sequence:
-//             a c d e f g h i j z
-//
-//   H = header   P = padding   a...z = keys  v = value | = block boundary
-//
-//   Encoded file:  H a cv dv ev P | H e ev fv gv hv P | H h hv iv jv z
-//   Decoded in blocks yields:
-//           Block 1: range [a, e) with kv pairs cv, dv
-//           Block 2: range [e, h) with kv pairs ev, fv, gv
-//           Block 3: range [h, z) with kv pairs hv, iv, jv
-//
-//   NOTE: All blocks except for the final block will have one last
-//   value which will not be used.  This isn't actually a waste since
-//   if the next KV pair wouldn't fit within the block after the value
-//   then the space after the final key to the next 1MB boundary would
-//   just be padding anyway.
-struct RangeFileWriter : public IRangeFileWriter {
-	explicit RangeFileWriter(Reference<IBackupFile> file = Reference<IBackupFile>(), int blockSize = 0)
-	  : file(file), blockSize(blockSize), blockEnd(0), fileVersion(BACKUP_AGENT_SNAPSHOT_FILE_VERSION) {}
-
-	// Handles the first block and internal blocks.  Ends current block if needed.
-	// The final flag is used in simulation to pad the file's final block to a whole block size
-	static Future<Void> newBlock(RangeFileWriter* self, int bytesNeeded, bool final = false) {
-		// Write padding to finish current block if needed
-		int bytesLeft = self->blockEnd - self->file->size();
-		if (bytesLeft > 0) {
-			Value paddingFFs = makePadding(bytesLeft);
-			co_await self->file->append(paddingFFs.begin(), bytesLeft);
-		}
-
-		if (final) {
-			ASSERT(g_network->isSimulated());
-			co_return;
-		}
-
-		// Set new blockEnd
-		self->blockEnd += self->blockSize;
-
-		// write Header
-		co_await self->file->append((uint8_t*)&self->fileVersion, sizeof(self->fileVersion));
-
-		// If this is NOT the first block then write duplicate stuff needed from last block
-		if (self->blockEnd > self->blockSize) {
-			co_await self->file->appendStringRefWithLen(self->lastKey);
-			co_await self->file->appendStringRefWithLen(self->lastKey);
-			co_await self->file->appendStringRefWithLen(self->lastValue);
-		}
-
-		// There must now be room in the current block for bytesNeeded or the block size is too small
-		if (self->file->size() + bytesNeeded > self->blockEnd)
-			throw backup_bad_block_size();
-
-		co_return;
-	}
-
-	// Used in simulation only to create backup file sizes which are an integer multiple of the block size
-	Future<Void> padEnd(bool final) override {
-		ASSERT(g_network->isSimulated());
-		if (file->size() > 0) {
-			return newBlock(this, 0, final);
-		}
-		return Void();
-	}
-
-	// Ends the current block if necessary based on bytesNeeded.
-	Future<Void> newBlockIfNeeded(int bytesNeeded) {
-		if (file->size() + bytesNeeded > blockEnd)
-			return newBlock(this, bytesNeeded);
-		return Void();
-	}
-
-	// Start a new block if needed, then write the key and value
-	static Future<Void> writeKV_impl(RangeFileWriter* self, Key k, Value v) {
-		int toWrite = sizeof(int32_t) + k.size() + sizeof(int32_t) + v.size();
-		co_await self->newBlockIfNeeded(toWrite);
-		co_await self->file->appendStringRefWithLen(k);
-		co_await self->file->appendStringRefWithLen(v);
-		self->lastKey = k;
-		self->lastValue = v;
-		co_return;
-	}
-
-	Future<Void> writeKV(Key k, Value v) override { return writeKV_impl(this, k, v); }
-
-	// Write begin key or end key.
-	static Future<Void> writeKey_impl(RangeFileWriter* self, Key k) {
-		int toWrite = sizeof(uint32_t) + k.size();
-		co_await self->newBlockIfNeeded(toWrite);
-		co_await self->file->appendStringRefWithLen(k);
-		co_return;
-	}
-
-	Future<Void> writeKey(Key k) override { return writeKey_impl(this, k); }
-
-	Future<Void> finish() override { return Void(); }
-
-	Reference<IBackupFile> file;
-	int blockSize;
-
-private:
-	int64_t blockEnd;
-	uint32_t fileVersion;
-	Key lastKey;
-	Key lastValue;
-};
-
-void decodeKVPairs(StringRefReader* reader, Standalone<VectorRef<KeyValueRef>>* results) {
-	// Read begin key, if this fails then block was invalid.
-	uint32_t kLen = reader->consumeNetworkUInt32();
-	const uint8_t* k = reader->consume(kLen);
-	results->push_back(results->arena(), KeyValueRef(KeyRef(k, kLen), ValueRef()));
-	KeyRef prevKey = KeyRef(k, kLen);
-	// Read kv pairs and end key
-	while (1) {
-		// Read a key.
-		kLen = reader->consumeNetworkUInt32();
-		k = reader->consume(kLen);
-
-		// If eof reached or first value len byte is 0xFF then a valid block end was reached.
-		if (reader->eof() || *reader->rptr == 0xFF) {
-			results->push_back(results->arena(), KeyValueRef(KeyRef(k, kLen), ValueRef()));
-			break;
-		}
-
-		// Read a value, which must exist or the block is invalid
-		uint32_t vLen = reader->consumeNetworkUInt32();
-		const uint8_t* v = reader->consume(vLen);
-
-		results->push_back(results->arena(), KeyValueRef(KeyRef(k, kLen), ValueRef(v, vLen)));
-
-		// If eof reached or first byte of next key len is 0xFF then a valid block end was reached.
-		if (reader->eof() || *reader->rptr == 0xFF)
-			break;
-	}
-
-	// Make sure any remaining bytes in the block are 0xFF
-	for (auto b : reader->remainder())
-		if (b != 0xFF)
-			throw restore_corrupted_data_padding();
-}
-
 static Reference<IBackupContainer> getBackupContainerWithProxy(Reference<IBackupContainer> _bc) {
 	Reference<IBackupContainer> bc = IBackupContainer::openContainer(
 	    _bc->getURL(), fileBackupAgentProxy, _bc->getEncryptionKeyFileName(), _bc->getEncryptionBlockSize());
 	return bc;
-}
-
-Standalone<VectorRef<KeyValueRef>> decodeRangeFileBlock(const Standalone<StringRef>& buf) {
-	Standalone<VectorRef<KeyValueRef>> results({}, buf.arena());
-	StringRefReader reader(buf, restore_corrupted_data());
-
-	// Read header, currently only decoding BACKUP_AGENT_SNAPSHOT_FILE_VERSION
-	if (reader.consume<int32_t>() != BACKUP_AGENT_SNAPSHOT_FILE_VERSION)
-		throw restore_unsupported_file_version();
-
-	// Read begin key, if this fails then block was invalid.
-	uint32_t beginKeyLen = reader.consumeNetworkUInt32();
-	const uint8_t* beginKey = reader.consume(beginKeyLen);
-	results.push_back(results.arena(), KeyValueRef(KeyRef(beginKey, beginKeyLen), ValueRef()));
-
-	// Read kv pairs and end key
-	while (1) {
-		// If eof reached or first value len byte is 0xFF then a valid block end was reached.
-		if (reader.eof() || *reader.rptr == 0xFF) {
-			break;
-		}
-
-		// Read a key, which must exist or the block is invalid
-		uint32_t kLen = reader.consumeNetworkUInt32();
-		const uint8_t* k = reader.consume(kLen);
-
-		// If eof reached or first value len byte is 0xFF then a valid block end was reached.
-		if (reader.eof() || *reader.rptr == 0xFF) {
-			// The last block in the file, will have Read End key.
-			results.push_back(results.arena(), KeyValueRef(KeyRef(k, kLen), ValueRef()));
-			break;
-		}
-
-		// Read a value, which must exist or the block is invalid
-		uint32_t vLen = reader.consumeNetworkUInt32();
-		const uint8_t* v = reader.consume(vLen);
-		results.push_back(results.arena(), KeyValueRef(KeyRef(k, kLen), ValueRef(v, vLen)));
-	}
-
-	// Make sure any remaining bytes in the block are 0xFF
-	for (auto b : reader.remainder())
-		if (b != 0xFF)
-			throw restore_corrupted_data_padding();
-
-	return results;
-}
-
-Future<Standalone<VectorRef<KeyValueRef>>> decodeRangeFileBlock(Reference<IAsyncFile> file,
-                                                                int64_t offset,
-                                                                int len,
-                                                                Database cx) {
-	Standalone<StringRef> buf = makeString(len);
-	int rLen = co_await uncancellable(holdWhile(buf, file->read(mutateString(buf), len, offset)));
-	if (rLen != len)
-		throw restore_bad_read();
-
-	simulateBlobFailure();
-
-	Standalone<VectorRef<KeyValueRef>> results({}, buf.arena());
-	StringRefReader reader(buf, restore_corrupted_data());
-	Arena arena;
-	try {
-		int32_t file_version = reader.consume<int32_t>();
-		if (file_version != BACKUP_AGENT_SNAPSHOT_FILE_VERSION) {
-			throw restore_unsupported_file_version();
-		}
-		decodeKVPairs(&reader, &results);
-		co_return results;
-	} catch (Error& e) {
-		TraceEvent(SevWarn, "FileRestoreDecodeRangeFileBlockFailed")
-		    .error(e)
-		    .detail("Filename", file->getFilename())
-		    .detail("BlockOffset", offset)
-		    .detail("BlockLen", len)
-		    .detail("ErrorRelativeOffset", reader.rptr - buf.begin())
-		    .detail("ErrorAbsoluteOffset", reader.rptr - buf.begin() + offset);
-		throw;
-	}
-}
-
-// Very simple format compared to KeyRange files.
-// Header, [Key, Value]... Key len
-struct LogFileWriter {
-	explicit LogFileWriter(Reference<IBackupFile> file = Reference<IBackupFile>(), int blockSize = 0)
-	  : file(file), blockSize(blockSize), blockEnd(0) {}
-
-	// Start a new block if needed, then write the key and value
-	static Future<Void> writeKV_impl(LogFileWriter* self, Key k, Value v) {
-		// If key and value do not fit in this block, end it and start a new one
-		int toWrite = sizeof(int32_t) + k.size() + sizeof(int32_t) + v.size();
-		if (self->file->size() + toWrite > self->blockEnd) {
-			// Write padding if needed
-			int bytesLeft = self->blockEnd - self->file->size();
-			if (bytesLeft > 0) {
-				Value paddingFFs = makePadding(bytesLeft);
-				co_await self->file->append(paddingFFs.begin(), bytesLeft);
-			}
-
-			// Set new blockEnd
-			self->blockEnd += self->blockSize;
-
-			// write the block header
-			co_await self->file->append((uint8_t*)&BACKUP_AGENT_MLOG_VERSION, sizeof(BACKUP_AGENT_MLOG_VERSION));
-		}
-
-		co_await self->file->appendStringRefWithLen(k);
-		co_await self->file->appendStringRefWithLen(v);
-
-		// At this point we should be in whatever the current block is or the block size is too small
-		if (self->file->size() > self->blockEnd)
-			throw backup_bad_block_size();
-
-		co_return;
-	}
-
-	Future<Void> writeKV(Key k, Value v) { return writeKV_impl(this, k, v); }
-
-	Reference<IBackupFile> file;
-	int blockSize;
-
-private:
-	int64_t blockEnd;
-};
-
-// input: a string of [param1, param2], [param1, param2] ..., [param1, param2]
-// output: a vector of [param1, param2] after removing the length info
-Standalone<VectorRef<KeyValueRef>> decodeMutationLogFileBlock(const Standalone<StringRef>& buf) {
-	Standalone<VectorRef<KeyValueRef>> results({}, buf.arena());
-	StringRefReader reader(buf, restore_corrupted_data());
-
-	// Read header, currently only decoding version BACKUP_AGENT_MLOG_VERSION
-	if (reader.consume<int32_t>() != BACKUP_AGENT_MLOG_VERSION)
-		throw restore_unsupported_file_version();
-
-	// Read k/v pairs.  Block ends either at end of last value exactly or with 0xFF as first key len byte.
-	while (1) {
-		// If eof reached or first key len bytes is 0xFF then end of block was reached.
-		if (reader.eof() || *reader.rptr == 0xFF)
-			break;
-
-		// Read key and value.  If anything throws then there is a problem.
-		uint32_t kLen = reader.consumeNetworkUInt32();
-		const uint8_t* k = reader.consume(kLen);
-		uint32_t vLen = reader.consumeNetworkUInt32();
-		const uint8_t* v = reader.consume(vLen);
-
-		results.push_back(results.arena(), KeyValueRef(KeyRef(k, kLen), ValueRef(v, vLen)));
-	}
-
-	// Make sure any remaining bytes in the block are 0xFF
-	for (auto b : reader.remainder())
-		if (b != 0xFF)
-			throw restore_corrupted_data_padding();
-
-	return results;
-}
-
-Future<Standalone<VectorRef<KeyValueRef>>> decodeMutationLogFileBlock(Reference<IAsyncFile> file,
-                                                                      int64_t offset,
-                                                                      int len) {
-	Standalone<StringRef> buf = makeString(len);
-	int rLen = co_await file->read(mutateString(buf), len, offset);
-	if (rLen != len)
-		throw restore_bad_read();
-
-	try {
-		co_return decodeMutationLogFileBlock(buf);
-	} catch (Error& e) {
-		TraceEvent(SevWarn, "FileRestoreCorruptLogFileBlock")
-		    .error(e)
-		    .detail("Filename", file->getFilename())
-		    .detail("BlockOffset", offset)
-		    .detail("BlockLen", len);
-		throw;
-	}
 }
 
 Future<Void> checkTaskVersion(Database cx, Reference<Task> task, StringRef name, uint32_t version) {
@@ -8836,19 +8480,4 @@ static Future<Void> writeKVs(Database cx, Standalone<VectorRef<KeyValueRef>> kvs
 	}
 
 	TraceEvent(SevFRTestInfo, "TransformDatabaseContentsWriteKVDone").detail("Begin", begin).detail("End", end);
-}
-
-void simulateBlobFailure() {
-	if (buggify() && deterministicRandom()->random01() < 0.01) { // Simulate blob failures
-		double i = deterministicRandom()->random01();
-		if (i < 0.5) {
-			throw http_request_failed();
-		} else if (i < 0.7) {
-			throw connection_failed();
-		} else if (i < 0.8) {
-			throw timed_out();
-		} else if (i < 0.9) {
-			throw lookup_failed();
-		}
-	}
 }
